+++ apps/exams/serializers/__init__.py
@@ -1,16 +1,16 @@
 from .admin.submissions_list import AdminExamSubmissionListResponseSerializer
-from .error_serializers import ErrorDetailSerializer, ErrorResponseSerializer
-from .exam_submission import (
+from .error_serializers import ErrorResponseSerializer
+from .student.submissions_create import (
     ExamSubmissionCreateResponseSerializer,
     ExamSubmissionCreateSerializer,
 )
-from .student.status import ExamStatusResponseSerializer
-from .student.submit_exam import (
+from .student.deployments_status import ExamStatusResponseSerializer
+from .student.submissions_submit import (
     AnswerItemSerializer,
     SubmitExamRequestSerializer,
     SubmitExamResponseSerializer,
 )
-from .student.take_exam import (
+from .student.deployments_take import (
     CheckCodeRequestSerializer,
     TakeExamRequestSerializer,
     TakeExamResponseSerializer,
@@ -28,5 +28,4 @@ __all__ = [
     "ExamSubmissionCreateSerializer",
     "ExamSubmissionCreateResponseSerializer",
     "ErrorResponseSerializer",
-    "ErrorDetailSerializer",
 ]
