+++ apps/exams/serializers/error_serializers.py
@@ -8,8 +8,3 @@ class ErrorResponseSerializer(serializers.Serializer[Any]):
 
     error_detail = serializers.CharField()
 
-
-class ErrorDetailSerializer(serializers.Serializer[Any]):
-    """권한/인증 에러 응답 스키마."""
-
-    detail = serializers.CharField()
