+++ apps/exams/serializers/student/exam_status.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from apps.exams.serializers.student.status import ExamStatusResponseSerializer
+from apps.exams.serializers.student.deployments_status import ExamStatusResponseSerializer
 
 __all__ = ["ExamStatusResponseSerializer"]
