+++ apps/exams/services/__init__.py
@@ -1,5 +1,5 @@
-from .student.submit_exam import submit_exam
-from .student.take_exam import build_take_exam_response, take_exam
+from .student.submissions_submit import submit_exam
+from .student.deployments_take import build_take_exam_response, take_exam
 
 __all__ = [
     "build_take_exam_response",
