+++ apps/exams/urls/student.py
@@ -1,10 +1,10 @@
 from django.urls import path
 
 from apps.exams.views import CheckCodeAPIView, ExamListView, TakeExamAPIView
-from apps.exams.views.exam_result import ExamSubmissionDetailView
-from apps.exams.views.exam_submission import ExamSubmissionCreateAPIView
-from apps.exams.views.student.cheating import ExamCheatingUpdateAPIView
-from apps.exams.views.student.status import ExamStatusCheckAPIView
+from apps.exams.views.student.submissions_result import ExamSubmissionDetailView
+from apps.exams.views.student.submissions_create import ExamSubmissionCreateAPIView
+from apps.exams.views.student.deployments_cheating import ExamCheatingUpdateAPIView
+from apps.exams.views.student.deployments_status import ExamStatusCheckAPIView
 
 app_name = "exams"
