+++ apps/exams/views/admin/exams_create.py
@@ -13,10 +13,7 @@ from apps.exams.serializers.admin.exams_create import (
     AdminExamCreateRequestSerializer,
     AdminExamCreateResponseSerializer,
 )
-from apps.exams.serializers.error_serializers import (
-    ErrorDetailSerializer,
-    ErrorResponseSerializer,
-)
+from apps.exams.serializers.error_serializers import ErrorResponseSerializer
 from apps.exams.services.admin.exams_create import (
     ExamCreateConflictError,
     ExamCreateNotFoundError,
@@ -41,8 +38,8 @@ class AdminExamCreateAPIView(APIView):
         responses={
             201: AdminExamCreateResponseSerializer,
             400: OpenApiResponse(ErrorResponseSerializer, description="유효하지 않은 시험 생성 요청"),
-            401: OpenApiResponse(ErrorDetailSerializer, description="인증 실패"),
-            403: OpenApiResponse(ErrorDetailSerializer, description="쪽지시험 생성 권한 없음"),
+            401: OpenApiResponse(ErrorResponseSerializer, description="인증 실패"),
+            403: OpenApiResponse(ErrorResponseSerializer, description="쪽지시험 생성 권한 없음"),
             404: OpenApiResponse(ErrorResponseSerializer, description="과목 정보 없음"),
             409: OpenApiResponse(ErrorResponseSerializer, description="동일한 이름의 시험 존재"),
         },
