+++ apps/exams/views/student/exam_status.py
@@ -1,5 +1,5 @@
 from __future__ import annotations
 
-from apps.exams.views.student.status import ExamStatusCheckAPIView
+from apps.exams.views.student.deployments_status import ExamStatusCheckAPIView
 
 ExamStatusAPIView = ExamStatusCheckAPIView
